+++ FE/src/components/common/tabs/tab/TabCompound/compounds/TabList.tsx
@@ -1,16 +1,16 @@
-import classNames from "classnames";
-import { PropsWithChildren } from "react";
-import { useTab } from "../TabCompound";
+// import classNames from "classnames";
+// import { PropsWithChildren } from "react";
+// import { useTab } from "../TabCompound";
 
-const TabList = ({ children }: PropsWithChildren) => {
-  const { align } = useTab<string>();
+// const TabList = ({ children }: PropsWithChildren) => {
+//   const { align } = useTab<string>();
 
-  const tabStyle = classNames(
-    tabAlign[align],
-    "w-full overflow-x-scroll scrollbar-hide",
-  );
+//   const tabStyle = classNames(
+//     tabAlign[align],
+//     "w-full overflow-x-scroll scrollbar-hide",
+//   );
 
-  return <div className={tabStyle}>{children}</div>;
-};
+//   return <div className={tabStyle}>{children}</div>;
+// };
 
-export default TabList;
+// export default TabList;
