+++ FE/src/components/programCreate/Participant.tsx
@@ -1,11 +1,7 @@
 "use client";
 
-// import MemberTableWrapper from "../common/memberTable/MemberTableWrapper";
-// import TabAsChild from "../common/tabs/TabAsChild/TabAsChild";
-// // import ManageTable from "../manage/member/memberManageTable/ManageTable";
-// import ParticipantTable from "./participantTable/ParticipantTable";
-// import ACTIVE_STATUS from "@/constants/ACTIVE_STATUS";
-// import { ActiveStatusWithAll } from "@/types/member";
+import MemberActiveStatusTab from "../common/tabs/MemberActiveStatusTab";
+import ParticipantTable from "./participantTable/ParticipantTable";
 
 interface MemberTableProps {
   members: Set<number>;
@@ -13,31 +9,26 @@ interface MemberTableProps {
   onClickHeaderCheckBox?: (selected: boolean) => void;
 }
 
-const Participant = ({} // members,
-// setMembers,
-// onClickHeaderCheckBox,
-: MemberTableProps) => {
+const Participant = ({
+  members,
+  setMembers,
+  onClickHeaderCheckBox,
+}: MemberTableProps) => {
   return (
-    // <TabAsChild<ActiveStatusWithAll>
-    //   defaultSelected="all"
-    //   options={Object.values(ACTIVE_STATUS.TAB_WITH_ALL)}
-    //   tabSize="lg"
-    //   baseColor="gray"
-    //   pointColor="teal"
-    //   align="line"
-    // >
-    //   {({ selectedItem }) => (
-    //     <div className="mt-4">
-    //       <ParticipantTable
-    //         members={members}
-    //         setMembers={setMembers}
-    //         selectedActive={selectedItem}
-    //         onClickHeaderCheckBox={onClickHeaderCheckBox}
-    //       />
-    //     </div>
-    //   )}
-    // </TabAsChild>
-    <></>
+    <section>
+      <MemberActiveStatusTab>
+        {(selectedItem) => (
+          <div className="mt-6">
+            <ParticipantTable
+              members={members}
+              setMembers={setMembers}
+              selectedActive={selectedItem}
+              onClickHeaderCheckBox={onClickHeaderCheckBox}
+            />
+          </div>
+        )}
+      </MemberActiveStatusTab>
+    </section>
   );
 };
